refactor: split app state tool executors
